% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{abis_signature}
\alias{abis_signature}
\title{This is the signature matrix generated by tha abis authors}
\format{
An object of class \code{data.frame} with 1296 rows and 17 columns.
}
\usage{
abis_signature
}
\description{
It is internally used by run_abis() but can be exported to allow modifications.
}
\references{
\url{https://github.com/giannimonaco/ABIS}
}
\keyword{datasets}
