% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconfectHelpers-package.R
\docType{package}
\name{deconfectHelpers-package}
\alias{deconfectHelpers}
\alias{deconfectHelpers-package}
\title{deconfectHelpers: Hopefully Helpful}
\description{
A loose collection of helper functions for my MD-project.
}
\author{
\strong{Maintainer}: Jonathan Speh \email{michel-jonathan.speh@charite.com} (\href{https://orcid.org/0000-0002-6204-4610}{ORCID})

}
\keyword{internal}
