% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{find_suspicious_names}
\alias{find_suspicious_names}
\title{Find gene names that excel converte to dates}
\usage{
find_suspicious_names(dat, gene_names, index = TRUE)
}
\arguments{
\item{dat}{A data frame containing gene names as a column.}

\item{gene_names}{A string specifying the column to check.}

\item{index}{A bool indicateing wether to return the column index of suspicious genes. Default is TRUE}
}
\value{
A tibble containing the suspicious genes
}
\description{
Find gene names that excel converte to dates
}
\examples{
2 + 2
}
