% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_deconvolution.R
\name{fix_cell_names_legacy}
\alias{fix_cell_names_legacy}
\title{Funktion to unify the names of immunecells. Should be replaced}
\usage{
fix_cell_names_legacy(cell_names)
}
\arguments{
\item{cell_names}{A vector of immune cell names}
}
\value{
A vector of standardised cell names
}
\description{
Funktion to unify the names of immunecells. Should be replaced
}
\examples{
fix_cell_names(c("NK","B.Naive", "B.Memory"), method = "abis")

}
