% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cell_hirarchy.R
\name{get_cumulative_proportions}
\alias{get_cumulative_proportions}
\title{A function to get cumulative proportions of cell types in a hierarchic tree}
\usage{
get_cumulative_proportions(node, return_all = FALSE)
}
\arguments{
\item{node}{A data.tree object with cell types and samples as attributes}

\item{return_all}{A boolean to specify whether original and cumulative proportions. Appends _cumulative to samples if true}
}
\value{
A dataframe with one row per cell type and columns for samples. Contains two columns per sample if return_all is TRUE
}
\description{
A function to get cumulative proportions of cell types in a hierarchic tree
}
\examples{
\dontrun{
get_cumulative_proportions(cell_tree)
}
}
