% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{get_rmse}
\alias{get_rmse}
\title{Compute rmse of deconvolution}
\usage{
get_rmse(deconvolution, actual)
}
\arguments{
\item{deconvolution}{deconvolution results in long format}

\item{actual}{actual fractions in long format}
}
\value{
rmse as numeric value
}
\description{
Compute rmse of deconvolution
}
\examples{
\dontrun{
get_rmse(deconvolution, actual)
}
}
