% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cell_hirarchy.R
\name{make_cell_tree}
\alias{make_cell_tree}
\title{Create a hierarchic data tree from deconvolution results}
\usage{
make_cell_tree(cell_mix)
}
\arguments{
\item{cell_mix}{A data frame containing sample, cell_type and proportion as columns.}
}
\value{
A data.tree object were each sample is one attribute
}
\description{
Create a hierarchic data tree from deconvolution results
}
\examples{
\dontrun{
maxe_cell_tree(mixture)
}

}
