% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deconvolution.R
\name{plot_deconvolution_corrs}
\alias{plot_deconvolution_corrs}
\title{Show corrolation between deconvolution results and actual fractions}
\usage{
plot_deconvolution_corrs(
  deconvolution,
  actual,
  add_x_y_line = TRUE,
  add_metrics = TRUE,
  remove_missing_celltypes = TRUE
)
}
\arguments{
\item{deconvolution}{deconvolution results in long format}

\item{actual}{actual fractions in long format}

\item{add_x_y_line}{Boolean to specify wether to plot red x == y line}

\item{add_metrics}{Boolean to specify wether to add RMSE and pearsons corrolation}

\item{remove_missing_celltypes}{A boolean weather to remove celltypes that are only in one of the datasets}
}
\value{
a gggplot2 object
}
\description{
Show corrolation between deconvolution results and actual fractions
}
\examples{
\dontrun{
plot_deconvolution_corrs(deconvolution, actual)
}
}
