% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deconvolution.R
\name{plot_deconvolution_violin}
\alias{plot_deconvolution_violin}
\title{Plot Violin}
\usage{
plot_deconvolution_violin(deconvolution, actual)
}
\arguments{
\item{deconvolution}{deconvolution results in long format}

\item{actual}{actual fractions in long format}
}
\value{
A ggplot2 object
}
\description{
Plot Violin
}
\examples{
\dontrun{
plot_deconvolution_violin(deconvultion, actual)
}
}
