% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_deconvolution.R
\name{process_immunedeconv}
\alias{process_immunedeconv}
\title{Transpose the results of immunedeconv and turns into long format}
\usage{
process_immunedeconv(cell_mix)
}
\arguments{
\item{cell_mix}{The results of a immunedeconv deconvolution. Must contain columns for cell type and s row for each sample}
}
\value{
A dataframe containing sample, cell_type and proportions
}
\description{
Transpose the results of immunedeconv and turns into long format
}
\examples{
\dontrun{
process_immunedeconv(mixture)
}
}
