% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_calls.R
\name{run_abis}
\alias{run_abis}
\title{A function to call the abis deconvolution method}
\usage{
run_abis(mixture, use_build_in_signature = TRUE, custom_signature = NULL)
}
\arguments{
\item{mixture}{Bulkmatrix with rownames for each gene and colnames for each sample}

\item{use_build_in_signature}{Boolean weather to use the signature matrix provided by the abis authors}

\item{custom_signature}{An alternative signature to use. Needs to have genes as rownames and cell types as columns}
}
\value{
A Matrix containing the results of deconvolution
}
\description{
A function to call the abis deconvolution method
}
\examples{
\dontrun{
run_abis(mixture)
}
}
