% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{run_cellanneal_r}
\alias{run_cellanneal_r}
\title{run_cellanneal_r}
\usage{
run_cellanneal_r(
  signature,
  mixture,
  disp_min = 0.5,
  bulk_min = 1e-05,
  bulk_max = 0.01,
  maxiter = 1000
)
}
\arguments{
\item{signature}{Path to the signature csv-file}

\item{mixture}{Path to the mixture csv-file}

\item{disp_min}{Minimum sclaed dispersion, defaults to 0.5}

\item{bulk_min}{Minimum expression in mixture, defaults to 1e-5}

\item{bulk_max}{Maximum expression in mixture, defualts to 0.01}

\item{maxiter}{max number of iteration, defaults to 1000}
}
\value{
A dataframe containing the results of celltype deconvolution
}
\description{
Wrapper for cellaneals "make_gene_dictionary" and "deconvolve" funktions.
Takes paths of signature and mixture file and returns dataframe with fractions
cellanneal has been developed by Lisa Buchauer, see https://github.com/LiBuchauer/cellanneal for mor information
}
\examples{
\dontrun{
run_cellanneal_r(path_to_signature, path_to_mixture)
}
}
