% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_deconvolution.R
\name{fix_cell_names}
\alias{fix_cell_names}
\title{Funktion to unify the names of immunecells}
\usage{
fix_cell_names(cell_names)
}
\arguments{
\item{cell_names}{A vector of immune cell names}
}
\value{
A vector of standardised cell names
}
\description{
Funktion to unify the names of immunecells
}
\examples{
fix_cell_names(c("bcells", "T cells CD4", "T reg"))

}
