% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deconvolution.R
\name{plot_deconvolution_celltype_corrs}
\alias{plot_deconvolution_celltype_corrs}
\title{Scatterplot with fitline for deconvolution results}
\usage{
plot_deconvolution_celltype_corrs(
  deconvolution,
  actual,
  add_x_y_line = TRUE,
  remove_missing_celltypes = TRUE
)
}
\arguments{
\item{deconvolution}{deconvolution results in long format}

\item{actual}{actual fractions in long format}

\item{add_x_y_line}{Boolean to specify wether to plot red x == y line}

\item{remove_missing_celltypes}{A boolean weather to remove celltypes that are only in one of the datasets}
}
\value{
A plot for each cell_type in the datasets
}
\description{
Scatterplot with fitline for deconvolution results
}
\examples{
\dontrun{
plot_deconvolution_celltype_corrs(deconvolution, actual)
}
}
