% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deconvolution.R
\name{plot_deconvolution_celltype_corrs}
\alias{plot_deconvolution_celltype_corrs}
\title{Scatterplot with fitline for deconvolution results}
\usage{
plot_deconvolution_celltype_corrs(deconvolution, actual)
}
\arguments{
\item{deconvolution}{deconvolution results in long format}

\item{actual}{actual fractions in long format}
}
\value{
A plot for each cell_type in the datasets
}
\description{
Scatterplot with fitline for deconvolution results
}
\examples{
#' \dontrun{
plot_deconvolution_celltype_corrs(deconvultion, actual)
}
}
