% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_deconvolution.R
\name{process_cellanneal}
\alias{process_cellanneal}
\title{Process cellanneal output}
\usage{
process_cellanneal(cell_mix)
}
\arguments{
\item{cell_mix}{}
}
\value{
deconvolution in long format
}
\description{
Process cellanneal output
}
\examples{
2 + 2
}
