% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_deconvolution.R
\name{process_immunedeconv}
\alias{process_immunedeconv}
\title{Transpose the results of immunedeconv and turns into long format}
\usage{
process_immunedeconv(cell_mix, fix_cell_names = TRUE)
}
\arguments{
\item{cell_mix}{The results of a immunedeconv deconvolution. Must contain columns for cell type and each sample}

\item{fix_cell_names}{A boolean to specify if the function should attempt to fix the cell name}
}
\value{
A dataframe containing sample, cell_type and proportopns
}
\description{
Transpose the results of immunedeconv and turns into long format
}
\examples{
\dontrun{
process_immunedeconv(mixture, fix_cell_names = TRUE)
}
}
