% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simulations.R
\name{simulate_sc}
\alias{simulate_sc}
\title{Helper to create a simulated sc-set including annotation}
\usage{
simulate_sc(
  n_genes = 10000,
  n_cells_est = 300,
  make_signature = TRUE,
  signature_path = "./signature.csv"
)
}
\arguments{
\item{n_genes}{number of genes in dataset}

\item{n_cells_est}{estimated number of cells, will only be exact if multiple of six}

\item{make_signature}{boolean weather to make a signature and write it to disk}

\item{signature_path}{path for signature matrix}
}
\value{
a list containing the counts and annotaiton
}
\description{
Helper to create a simulated sc-set including annotation
}
\examples{
simulate_sc(n_genes = 100, n_cells_est = 30, make_signature = FALSE)
}
